Move add_docs_params
