+++ src/marqo/core/unstructured_vespa_index/unstructured_add_document_handler.py
@@ -8,7 +8,7 @@ from marqo.api import exceptions as api_errors
 from marqo.core import constants
 from marqo.core.constants import MARQO_DOC_ID
 from marqo.core.document.add_documents_handler import AddDocumentsHandler, AddDocumentsError
-from marqo.core.document.models.add_docs_params import AddDocsParams
+from marqo.core.models.add_docs_params import AddDocsParams
 from marqo.core.document.tensor_fields_container import TensorFieldsContainer, MODALITY_FIELD_TYPE_MAP
 from marqo.core.models import UnstructuredMarqoIndex
 from marqo.core.models.marqo_index import FieldType
