+++ src/marqo/tensor_search/web/api_utils.py
@@ -4,7 +4,7 @@ from typing import Union, List, Optional, Dict
 
 from marqo.api.exceptions import InvalidArgError
 from marqo.tensor_search import enums
-from marqo.tensor_search.models.add_docs_objects import AddDocsParams, AddDocsBodyParams
+from marqo.tensor_search.models.add_docs_objects import AddDocsBodyParams
 from marqo.tensor_search.models.add_docs_objects import ModelAuth
