+++ tests/core/index_management/test_index_management.py
@@ -30,6 +30,9 @@ from marqo.core.models.marqo_index_request import FieldRequest
 from marqo.core.vespa_schema import for_marqo_index_request as vespa_schema_factory
 from marqo.s2_inference.s2_inference import get_model_properties_from_registry
 from marqo.vespa.exceptions import VespaActivationConflictError
+from marqo.tensor_search import tensor_search
+from marqo.core.models.add_docs_params import AddDocsParams
+from marqo.vespa.exceptions import VespaStatusError
 from marqo.vespa.models import VespaDocument
 from tests.marqo_test import MarqoTestCase
