+++ tests/core/monitoring/test_monitoring.py
@@ -10,7 +10,7 @@ from marqo.core.models.marqo_index import FieldType, FieldFeature, TextPreProces
 from marqo.core.models.marqo_index_request import FieldRequest
 from marqo.core.models.marqo_index_stats import MarqoIndexStats, VespaStats
 from marqo.tensor_search import tensor_search
-from marqo.tensor_search.models.add_docs_objects import AddDocsParams
+from marqo.core.models.add_docs_params import AddDocsParams
 from tests.marqo_test import MarqoTestCase, TestImageUrls
