+++ tests/s2_inference/test_generic_clip_model.py
@@ -2,7 +2,7 @@ import unittest
 
 import numpy as np
 import os
-from marqo.tensor_search.models.add_docs_objects import AddDocsParams
+from marqo.core.models.add_docs_params import AddDocsParams
 from marqo.api.exceptions import IndexNotFoundError
 from marqo.s2_inference.errors import UnknownModelError, ModelLoadError
 from marqo.tensor_search import tensor_search
