+++ tests/tensor_search/integ_tests/test_add_documents_combined.py
@@ -19,7 +19,7 @@ from marqo.core.models.marqo_index_request import FieldRequest
 from marqo.s2_inference import types
 from marqo.tensor_search import add_docs
 from marqo.tensor_search import tensor_search
-from marqo.tensor_search.models.add_docs_objects import AddDocsParams
+from marqo.core.models.add_docs_params import AddDocsParams
 from tests.marqo_test import MarqoTestCase, TestImageUrls
 from marqo.s2_inference.multimodal_model_load import infer_modality, Modality
 from marqo.tensor_search import streaming_media_processor
