+++ tests/tensor_search/integ_tests/test_add_documents_unstructured.py
@@ -13,7 +13,7 @@ from marqo.core.models.marqo_index import *
 from marqo.tensor_search import add_docs
 from marqo.tensor_search import enums
 from marqo.tensor_search import tensor_search
-from marqo.tensor_search.models.add_docs_objects import AddDocsParams
+from marqo.core.models.add_docs_params import AddDocsParams
 from tests.marqo_test import MarqoTestCase, TestImageUrls
