+++ tests/tensor_search/integ_tests/test_embed.py
@@ -1,4 +1,4 @@
-from marqo.tensor_search.models.add_docs_objects import AddDocsParams
+from marqo.core.models.add_docs_params import AddDocsParams
 from marqo.tensor_search import tensor_search
 from marqo.tensor_search import enums
 from marqo.tensor_search.models.api_models import BulkSearchQuery, BulkSearchQueryEntity, ScoreModifierLists
