+++ tests/tensor_search/integ_tests/test_get_document.py
@@ -1,7 +1,7 @@
 import functools
 import unittest
 
-from marqo.tensor_search.models.add_docs_objects import AddDocsParams
+from marqo.core.models.add_docs_params import AddDocsParams
 from marqo.tensor_search import enums
 from marqo.api.exceptions import IndexNotFoundError, InvalidDocumentIdError
 from marqo.tensor_search import tensor_search
