+++ tests/tensor_search/integ_tests/test_get_documents_by_ids.py
@@ -12,7 +12,7 @@ from marqo.api.exceptions import (
 )
 from marqo.tensor_search import enums
 from marqo.tensor_search import tensor_search
-from marqo.tensor_search.models.add_docs_objects import AddDocsParams
+from marqo.core.models.add_docs_params import AddDocsParams
 from tests.marqo_test import MarqoTestCase
 from marqo.core.models.marqo_index import *
 from marqo.core.models.marqo_index_request import FieldRequest
