+++ tests/tensor_search/integ_tests/test_hybrid_search.py
@@ -9,7 +9,7 @@ from marqo.core.models.hybrid_parameters import RetrievalMethod, RankingMethod, 
 from marqo.core.structured_vespa_index import common
 from marqo.tensor_search import tensor_search
 from marqo.tensor_search.enums import SearchMethod
-from marqo.tensor_search.models.add_docs_objects import AddDocsParams
+from marqo.core.models.add_docs_params import AddDocsParams
 from tests.marqo_test import MarqoTestCase, TestImageUrls
 from marqo import exceptions as base_exceptions
 import unittest
